requirements.txt
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
langchain
langchain-community
langchain-core
langgraph
langgraph-supervisor
python-dotenv
requests
fastapi
flask
streamlit
openai 
groq
langchain-groq 
langchain-tavily
uvicorn
langchain-openai